[FE: REFACTOR] Refactor typo
