+++ f_end/src/components/sensor/SensorChart.vue
@@ -5,9 +5,8 @@
 <script setup lang='ts'>
 import { defineProps, onMounted, ref, watch } from 'vue';
 import { Chart, registerables } from 'chart.js';
-import { useCensorData } from '@/store/censor';
+import { useSensorData } from '@/store/censor';
 
-const sensorData = useCensorData();
 const props = defineProps({
   topic: {
     type: String,
@@ -17,10 +16,10 @@ const props = defineProps({
 Chart.register(...registerables);
 const chart = ref(null);
 let chartInstance = null;
+const { data } = useSensorData();
 
 onMounted(() => {
   const ctx = document.getElementById(props.topic).getContext('2d');
-  const data = sensorData.getAllData();
   chartInstance = new Chart(ctx, {
     type: 'line',
     data: {
@@ -28,7 +27,7 @@ onMounted(() => {
       datasets: [
         {
           label: props.topic,
-          data: data.get(props.topic),
+          data: data[props.topic],
           fill: false,
           borderColor: 'rgb(75, 192, 192)',
           backgroundColor: 'rgb(255, 255, 255)',
@@ -37,9 +36,10 @@ onMounted(() => {
       ]
     }
   })
+  console.log(props.topic);
 });
 
-watch(() => sensorData.getAllData().get(props.topic), (newData) => {
+watch(() => data[props.topic], (newData) => {
   if (chartInstance) {
     chartInstance.data.labels = newData.map((_, idx) => (
       idx - newData.length + 1) * 5 + ' s'
@@ -47,7 +47,7 @@ watch(() => sensorData.getAllData().get(props.topic), (newData) => {
     chartInstance.data.datasets[0].data = newData;
     chartInstance.update();
   }
-}, {deep: true})
+}, { deep: true })
 </script>
 
 <style scoped>
